restructured anna command
